+++ laktory/spark/dataframe/__init__.py
@@ -40,3 +40,12 @@ def has_column(df: DataFrame, col: str) -> bool:
 # DataFrame Extensions
 DataFrame.schema_flat = schema_flat
 DataFrame.has_column = has_column
+
+# Spark Connect DataFrame Extensions
+try:
+    from pyspark.sql.connect.dataframe import DataFrame
+    DataFrame.schema_flat = schema_flat
+    DataFrame.has_column = has_column
+except ModuleNotFoundError:
+    pass
+